requirements.txt
Django>=5.0,<6.0
Pillow>=10.0.0
django-ckeditor>=6.7.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
requests>=2.31.0
django-environ>=0.11.0
gunicorn>=21.0.0
whitenoise>=6.6.0
